Fix type conversion warnings in PBGitGrapher.mm
Fixed three type conversion warnings where size_type (unsigned long)
was being implicitly converted to int by adding explicit casts:
- Line 183: add_line call with currentLanes->size()
- Line 204: numColumns assignment with currentLanes->size() - 1
- Line 206: numColumns assignment with currentLanes->size()

These changes are safe as the vector sizes in this context are
expected to be small and well within int range.

+++ Classes/git/PBGitGrapher.mm
@@ -180,7 +180,7 @@ void add_line(struct PBGitGraphLine *lines, int *nLines, int upper, int from, in
 			addedParent = YES;
 			PBGitLane *newLane = new PBGitLane(_laneIndex++, parentOID);
 			currentLanes->push_back(newLane);
-			add_line(lines, &currentLine, 0, currentLanes->size(), newPos, newLane->index());
+			add_line(lines, &currentLine, 0, (int)currentLanes->size(), newPos, newLane->index());
 		} @catch (NSException *exception) {
 		}
 	}
@@ -201,9 +201,9 @@ void add_line(struct PBGitGraphLine *lines, int *nLines, int upper, int from, in
 
 	// If a parent was added, we have room to not indent.
 	if (addedParent)
-		self.previous.numColumns = currentLanes->size() - 1;
+		self.previous.numColumns = (int)currentLanes->size() - 1;
 	else
-		self.previous.numColumns = currentLanes->size();
+		self.previous.numColumns = (int)currentLanes->size();
 
 	// Update the current lane to point to the new parent
 	if (currentLane) {
